+++ pioreactor/actions/leader/export_experiment_data.py
@@ -4,6 +4,7 @@
 from __future__ import annotations
 
 import os
+import re
 from contextlib import closing
 from contextlib import ExitStack
 from datetime import datetime
@@ -12,6 +13,10 @@ import click
 
 from pioreactor.config import config
 from pioreactor.logging import create_logger
+
+
+def is_valid_table_name(table_name: str) -> bool:
+    return bool(re.fullmatch(r"^[a-zA-Z]\w*$", table_name))
 
 
 def source_exists(cursor, table_name_to_check: str) -> bool:
@@ -30,10 +35,14 @@ def filter_to_timestamp_columns(column_names: list[str]) -> list[str]:
 
 
 def generate_timestamp_to_localtimestamp_clause(cursor, table_name: str) -> str:
-    # TODO: this assumes a timestamp column exists?
     columns = get_column_names(cursor, table_name)
     timestamp_columns = filter_to_timestamp_columns(columns)
+
+    if not timestamp_columns:
+        return ""
+
     clause = ",".join([f"datetime({c}, 'localtime') as {c}_localtime" for c in timestamp_columns])
+
     if clause:
         clause += ","
 
@@ -58,7 +67,7 @@ def export_experiment_data(
     logger = create_logger("export_experiment_data")
     logger.info(f"Starting export of table{'s' if len(tables) > 1 else ''}: {', '.join(tables)}.")
 
-    time = datetime.now().strftime("%Y%m%d%H%m%S")
+    time = datetime.now().strftime("%Y%m%d%H%M%S")
 
     with zipfile.ZipFile(output, mode="w", compression=zipfile.ZIP_DEFLATED) as zf, closing(
         sqlite3.connect(config["storage"]["database"])
@@ -68,33 +77,34 @@ def export_experiment_data(
 
             # so apparently, you can't parameterize the table name in python's sqlite3, so I
             # have to use string formatting (SQL-injection vector), but first check that the table exists (else fail)
-            if not source_exists(cursor, table):
+            if not (source_exists(cursor, table) and is_valid_table_name(table)):
                 raise ValueError(f"Table {table} does not exist.")
 
             timestamp_to_localtimestamp_clause = generate_timestamp_to_localtimestamp_clause(
                 cursor, table
             )
-            order_by = filter_to_timestamp_columns(
-                get_column_names(cursor, table)
-            ).pop()  # just take first...
 
-            if (
-                table == "pioreactor_unit_activity_data"
-                or table == "pioreactor_unit_activity_data_rollup"
-            ):
-                _partition_by_unit = True
+            timestamp_columns = filter_to_timestamp_columns(get_column_names(cursor, table))
+            if not timestamp_columns:
+                order_by = (
+                    "rowid"  # yes this is stupid, but I need a placeholder for the queries below
+                )
             else:
-                _partition_by_unit = partition_by_unit
+                order_by = timestamp_columns[0]  # just take first...
+
+            _partition_by_unit = partition_by_unit
+            if table in ("pioreactor_unit_activity_data", "pioreactor_unit_activity_data_rollup"):
+                _partition_by_unit = True
 
             if not _partition_by_unit:
                 if experiment is None:
-                    query = f"SELECT {timestamp_to_localtimestamp_clause} * from {table} ORDER BY :order_by"
-                    cursor.execute(query, {"order_by": order_by})
+                    query = f"SELECT {timestamp_to_localtimestamp_clause} * from {table} ORDER BY {order_by}"
+                    cursor.execute(query)
                     filename = f"{table}-{time}.csv"
 
                 else:
-                    query = f"SELECT {timestamp_to_localtimestamp_clause} * from {table} WHERE experiment=:experiment ORDER BY :order_by"
-                    cursor.execute(query, {"experiment": experiment, "order_by": order_by})
+                    query = f"SELECT {timestamp_to_localtimestamp_clause} * from {table} WHERE experiment=:experiment ORDER BY {order_by}"
+                    cursor.execute(query, {"experiment": experiment})
                     filename = f"{experiment}-{table}-{time}.csv"
 
                 filename = filename.replace(" ", "_")
@@ -108,6 +118,9 @@ def export_experiment_data(
                 os.remove(path_to_file)
 
             else:
+                if experiment is None:
+                    raise ValueError("Experiment name should be provided.")
+
                 query = f"SELECT {timestamp_to_localtimestamp_clause} * from {table} WHERE experiment=:experiment ORDER BY :order_by"
                 cursor.execute(query, {"experiment": experiment, "order_by": order_by})