+++ pioreactor/cli/pio.py
@@ -434,7 +434,7 @@ def get_tag_to_install(version_desired: Optional[str]) -> str:
 
 
 @update.command(name="app")
-@click.option("-b", "--branch", help="update to a branch on github")
+@click.option("-b", "--branch", help="install from a branch on github")
 @click.option("--source", help="use a URL or whl file")
 @click.option("-v", "--version", help="install a specific version, default is latest")
 def update_app(branch: Optional[str], source: Optional[str], version: Optional[str]) -> None:
@@ -469,8 +469,7 @@ def update_app(branch: Optional[str], source: Optional[str], version: Optional[s
         version_installed = release_metadata["tag_name"]
         for asset in release_metadata["assets"]:
             # add the following files to the release. They should ideally be idempotent!
-            # https://arslan.io/2019/07/03/how-to-write-idempotent-bash-scripts/
-            #
+
             # pre_update.sh runs (if exists)
             # `pip install pioreactor...whl` runs
             # update.sh runs (if exists)
@@ -508,7 +507,6 @@ def update_app(branch: Optional[str], source: Optional[str], version: Optional[s
                     ]
                 )
             elif asset_name == "post_update.sh":
-                # ex: post_update.sh can be used to restart machines
                 commands_and_priority.extend(
                     [
                         (f"wget -O /tmp/post_update.sh {url}", 99),
@@ -771,7 +769,7 @@ if whoami.am_I_leader():
             raise click.Abort()
 
     @update.command(name="ui")
-    @click.option("-b", "--branch", help="update to a branch on github")
+    @click.option("-b", "--branch", help="install from a branch on github")
     @click.option("--source", help="use a tar.gz file")
     @click.option("-v", "--version", help="install a specific version")
     def update_ui(branch: Optional[str], source: Optional[str], version: Optional[str]) -> None: