+++ pioreactor/whoami.py
@@ -108,7 +108,7 @@ def get_hashed_serial_number() -> str:
 def get_rpi_machine() -> str:
     if not is_testing_env():
         with open("/proc/device-tree/model") as f:
-            return f.read()
+            return f.read().strip()
     else:
         return "Raspberry Pi 3 - testing"
 
@@ -117,7 +117,7 @@ def get_image_git_hash() -> str:
     try:
         with open("/home/pioreactor/.pioreactor/.image_info") as f:
             return f.read().strip().split("=")[1]
-    except FileNotFoundError:
+    except OSError:  # catch FileNotFoundError, PermissionError, and other file-related exceptions
         return "<Failed to fetch>"