+++ CHANGELOG.md
@@ -1,8 +1,17 @@
-### 23.4.14q
+### 23.4.28
+ - improved detection of under-voltage, and power supply problems.
+ - pumps will halt if a MQTT disconnect occurs. This is to prevent the edge case when pumps are running on a worker, but not controllable from the UI due to an MQTT disconnect.
+ - improvements to backing up the SQLite3 database.
+ - improvements to self-test "Reference photodiode is correct magnitude"
+
+#### Bug fixes
+ - fix bug in `pio run od_calibration list`
+
+### 23.4.14
 
  - `pio update app` will default to installing the succeeding release of Pioreactor app, which may or may not be the latest. This is to ensure that no update script is skipped.
  - Added new "Past Experiments" page
- - Fix for "Reference photodiode is correct magnitude" self test.
+ - Fix for "Reference photodiode is correct magnitude" self-test.
 
 ### 23.4.4
  - Job growth_rate_calculating will dynamically choose initial values for its internal statistics.